common: Fix branch validation to match docs
The docs for flatpak_is_valid_branch() say branch names can't start with
a digit but the implementation doesn't enforce this, so change the
enforcement to disallow e.g. "99wip".

The implementation also disallows "." as the leading character for a
branch, so add that to the docs. I'm just guessing that "." was intended
to be disallowed but it makes sense; otherwise the file we create named
with the branch would be hidden.

Closes: #3023
Approved by: matthiasclasen